changed getIDFromAuth function in Utils.java to getUserAuthData where it gets id,mail and roles to be used later
added isAuthorizedForThis in Utiles.java
where it checks if user is admin or same as the one logged in for authentication purposes

+++ src/main/java/com/example/SpringBootMVC/post/service/PostServiceImpl.java
@@ -9,12 +9,14 @@ import com.example.SpringBootMVC.user.UserDetailsImpl;
 import com.example.SpringBootMVC.user.dto.UserResponse;
 import com.example.SpringBootMVC.user.entity.User;
 import com.example.SpringBootMVC.user.repository.UserRepository;
+import com.example.SpringBootMVC.utils.AuthUserDTO;
 import com.example.SpringBootMVC.utils.Utils;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -23,13 +25,15 @@ public class PostServiceImpl implements PostService {
     private PostRepository postRepository;
     private UserRepository userRepository;
     private Utils utils;
+
     public PostServiceImpl(PostRepository postRepository, UserRepository userRepository, Utils utils) {
         this.postRepository = postRepository;
         this.userRepository = userRepository;
         this.utils = utils;
     }
+
     @Override
-    public  List<PostResponse> getAllPosts() {
+    public List<PostResponse> getAllPosts() {
         return postRepository.findAll()
                 .stream()
                 .map(post -> new PostResponse(
@@ -63,11 +67,20 @@ public class PostServiceImpl implements PostService {
     @Override
     public String deletePost(Long id) {
 
-        Long authID = utils.getUserIDFromAuth();
         Post post = postRepository.findById(id).orElseThrow(() -> new RuntimeException("Post not found"));
-        /*User user = userRepository.findById(post.getUser().getId())
-                .orElseThrow(() -> new RuntimeException("User not found"));*/
-        postRepository.deleteById(post.getId());
-        return "Post Deleted successfully";
+        if (utils.isAuthorizedForThis(post.getUser().getId())) {
+            postRepository.deleteById(post.getId());
+            return "Post Deleted successfully";
+        } else {
+            return "User is not authorized to do this action";
+        }
+        /* AuthUserDTO authData = utils.getUserAuthData();*/
+
+       /* if (Objects.equals(authData.getId(), post.getUser().getId()) || authData.getRoles().contains("ROLE_ADMIN")) {
+            postRepository.deleteById(post.getId());
+            return "Post Deleted successfully";
+        }else {
+            return "User is not authorized to do this action";
+        }*/
     }
 }
